py/requirements.txt
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.64.1
idna==3.7
joblib==1.4.2
lxml==5.2.2
nltk==3.8.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.5
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
